Fix licm bug that could result in hoisting to the wrong loop
Fix how we insert hoisted expressions into the pre-headers of loops. LICM
mostly operates on basic-blocks, but when we hoist invariant expressions we
need to hoist out of the loop AST. Previously, to find the loop AST we would
look at the header basic-block, grab the first expression, and walk up the
tree until we found the parent loop. However, since basic blocks aren't updated
with updates to the AST, this would mean that if we hoisted the first expr, the
next thing to be hoisted would be doubly hoisted.

In reality this never occurred because the first expression was always a
defExpr and LICM was mistakenly not hoisting defExprs before. #6692 starts
hoisting defExprs, and this PR is needed to avoid the double hoisting. This PR
just creates a reference to the loopAST when the basic-block loop is first
created.