Remove dont_use_exit_use_chpl_exit_instead
The macro `define exit dont_use_exit_use_chpl_exit_instead` was used to prevent
accidental calls to exit instead of chpl_exit. Unfortunately this sort of macro
definition is error-prone because it replaces every `exit` token, not just
function calls. This is causing compilation failures with new versions of clang
on OSX with an error along the lines of:

```
 no member named 'dont_use_exit_use_chpl_exit_instead' in the global namespace
 using ::exit;
       ~~^
 ../../../../include/chplrt.h:41:17: note: expanded from macro 'exit'
 #define exit    dont_use_exit_use_chpl_exit_instead
```

This just removes the macro, and a later commit will take care of updating our
`lookForBadAlloc` script to check for calls to exit() as well. We want to avoid
calling exit so that we have a common place to set a breakpoint for debugging.