Updates based on some of Mike's feedback (more incoming)
Make the link to the initializers chip something that can be used multiple
times.  Rephrase "declares a type initializer method" to "declares an
initializer method on a type".  Explicitly mention compiler generated
initializers prior to the link to that section, to avoid confusion about why it
says we generate constructors and then link to compiler generated initializers.
Add an example of an ordinary pair of initializers to the original description.
Clarify the Initializer Body section by being a little bit more explicit about
the two phases and their interaction with the phase division indicator.  Add a
reference to the initializers chip at the end of the paragraph with an
indication that the rationale and additional details can be found there.  Add an
example of initializers calling parent initializers, and initializers calling
another initializer defined on the same type.