Restore performance for substring test
Restore performance for the "search within a string" test. #7371 fixed some
LICM reference bugs and as a result we're no longer hoisting some buffer temps
in String.this(r: range) (substring). Getting LICM to actually hoist these
temps again is a huge undertaking, so for now manually hoist a temp in the
String module. This is admittedly lame, but without this we pay a 250%
performance loss for the substring performance test.