compiler inserts deferred delete error into catch statements (#7387)
[reviewed by @mppf]

When a new error is thrown from a catch block, the old handled error is leaked. This bug is detailed in #7331.

To resolve the bug, the compiler changes it's inserted deletes for Errors. Instead of inserting the delete at the end of the catch block, it is inserted inside a DeferStmt at the beginning of the catch block. To ensure that an error is not deleted when it is thrown on, the compiler will insert checks to nil out the handled error variable before throws in catch blocks.