Increase the gasnet exit timeout under gasnet-{aries,gemini} again
Bump the default timeout again (first done in #7241). GASNet 1.30.0 added an
additional shutdown watchdog, which has caused us to see some sigalarms. We use
the multi-domain feature, which increases the shutdowm times so for now just
raise the default timeout. Note that there was no actual regression in the
shutdown times in 1.30.0, it's just that there is an extra watchdog whose limit
is too low when the multi-domain feature is being used.

For more info see https://github.com/chapel-lang/chapel/issues/7251 and
https://upc-bugs.lbl.gov/bugzilla/show_bug.cgi?id=3621