Initial update for 1.16.
Remove text that treats multi-ddata as the status quo for
locModel=numa.  Add Current Status sections for Single-ddata with
Separate Arrays and Multi-ddata, saying where they're at.

Still to do: the major thing is a rewrite of the descriptive stuff at
the end of the Summary.  I also need to re-read the Single-ddata with
Cyclic Localization section and reassure myself that (as I believe) the
prohibition against multiple mmap()ings referencing the same virtual
address is not applicable to it.