Remove references to pgi in using-a-cray doc
We don't build the module with support for pgi as a target compiler anymore,
and we don't trust pgi that much anyways, so don't "recommend" it as a backend
compiler.

Also remove the "change your host compiler section". We really recommend using
gnu, so don't let people know they can use anything else.