Remove notes about setting comm and tasking
For tasking we really want people to just default to qthreads. This might have
been more relevant when muxed was around, but now we really don't want fifo to
be used.

This also removes the note about setting the comm layer. ugni provides the best
performance, so we want users to use it if they're on a xe/xc. If they're on a
cs, gasnet is there only option, so there's no reason for them to change it.