Error doesn't include msg field
I don't think Error should include a .msg field - I think it
should be as close as possible to the minimum required to
be an "Error". (I.e. close to a virtual base class).

This PR removes the message field and updates the tests
that were using that behavior to instead use a StringError
from an ExampleErrors module.