Fix a race in our slurm-srun launcher while determining the slurm version
Fix a race in slurm-srun:determineSlurmVersion(). Previously we wrote the
contents of `sbatch --version` out to a tmp file and then read the file to
determine versions. Unfortunately, we used the same tmp file so this would
result in segfaults and other sporadic failures if 2 programs were running from
the same directory. While we run concurrent testing on crays, we use unique
subdirs so we haven't run into this in nightly testing. This was discovered by
the gasnet team's CI testing.

There's no reason to be using a tmp file anyways (this code was originally
copied from the slurm-gasnetrun_ibv launcher) so just use chpl_run_utility1K
instead. I copied the pattern we use in determineQsubVersion.

Note that slurm-gasnetrun_ibv and pbs-gasnetrun_ibv also need to be updated,
but I'll do that in a later PR. slurm-srun is one of our most commonly used
launchers, so I want to get this in a soon as possible.