Preserve comm=none correctness test
Since we can't add a .bad file for the hpl.hpcc2012.chpl test because
its output changes from run-to-run, it seems valuable to lock in
comm=none correctness testing.  This commit does so by adding a code
clone that shares the pre-existing comm-none.good file.  I tried to
make the .chpl file empty and use the same source as the original,
but couldn't get this to work for reasons that seemed too weird to
spend time on today (one day before feature freeze).