Don't depend on third-party source files to auto-rebuild a third-party lib
PR #7004 added code to automatically rebuild a third-party library if our
README/Makefiles changes or if any source file in the third-party dir changed.
Unfortunately depending on third-party sources had some unforeseen consequences
because of built-in make rules, so now only rebuild if our README/Makefiles
changes. We'll still auto-rebuild on version upgrades or if we bring in an
upstream patch or something so long as we're careful to update the README to
note version changes and mods.

The specific problem we ran into with depending on third-party sources was that
a built-in rule was causing us to auto-rebuild some gmp files that have a .y
file. Make will automatically rebuild a .c file if it has a .y file with a
newer timestamp than it. We could probably work around this, but it makes me
nervous about other built-in rules and I don't want to take any chances. For
gmp, this was effectively requiring a user to have yacc/bison installed in
order for `make check` to pass, which we do as part of the speculative build.

Also touch the specific .c files that were being rebuilt. They won't be rebuilt
by an implicit make rule anymore, but the gmp build might rebuild them, so make
sure that the generated .c has a newer timestamp than the source .y.