Fix an issue with returning a generic record created via a 'new' directly
Prior to this change, the following would fail with the error `illegal use of
function that does not return a value: 'init'` when the type defined an
initializer:

```chapel
proc Some(value: ?eltType) {
  return new Option(value);
}
```

This was due to an oversight in AST transformation.  I fixed a similar error for
concrete records with #6773 but didn't get back to the generic case until just
now.