Fix reduce intents for parallel iterators with forall.
This fixes the generated code for forall loops with
in intent and/or reduce intent, ex.:

    forall ... with (in somevar, + reduce anothervar) ...;

where the corresponding parallel iterator itself has a forall loop, ex.:

    iter myParallelIter(param tag: iterKind) {
      ...
      forall ... {
        yield ...;
      }
      ...
    }

Before this change, the same shadow variable for the above "somevar"
and "anothervar" would be accessed for all iterations of the forall
loop inside myParallelIter(). This can cause unintended data races for
reduce and 'in' forall intents. Cf. no data races for const forall intents
(cannot modify the shadow variables) and already-intended data race
for the ref forall intent.

This fix adds shadow variables for this forall loop as well.

This fix does not handle the case where the parallel iterator has
**nested** forall loops. Handling of that case should be straightforward.
I did not do it because I timed out and because it seems we run into
other/unrelated issues with nested forall loops in the parallel iterator.
I am leaving it for future work.

So if there are nested forall loops, the compiler now generates an error
when an 'in' or reduce intent is present. This is to avoid unintended
races in the generated code, given the state of implementation.

Also I am handling only reduce intents for the "old" implementation.
This is because in the old implementation the specific intent is not
easily discernible. The reduce intent is discernible only by the virtue
of the shadow variable being a subclass of the ReduceScanOp class.

Retiring BenH's original future and adding a more extensive test case.