Prevent certain data races upon 'in' forall intents
Before this change, the compiler would generate racy code
for a forall loop with an 'in'-intented shadow variable, ex.:

    forall ... with (in somevar) ...;

when the corresponding parallel iterator has a for loop over
another parallel iterator, ex.:

    iter myParallelIter(param tag: iterKind) {
      ...
      for anotherIter(iterKind.standalone) {
        yield ...;
      }
      ...
    }

The reason is that all iterations of the for-loop over anotherIter
would use the same shadow variable for 'somevar', allowing for
unintended data races.

The compiler changes to handle this situation properly would be too invasive
so I did not want to put them in so close to code freeze.
Leaving this as future work.

Instead, this change detects this situation and changes the 'in' intent
to 'const in' intent. So if 'somevar' is modified within the loop,
const-checking flags that as an error. If it is not modified
within the loop, there will be no data races. The user is warned
about the change from 'in' to 'const in'.

A couple of miniMD and stencil variants pass a tuple by 'in' intent.
Changing those to 'const in' intent.