Emphasize parens for deinit
With #7375, we require deinit functions to be declared with parens,
both as class/record deinitializers and module deinit functions.

Emphasize that in the spec by writing parens next to "deinit"
where they are discussed.
Emphasize that in CHIP 10 explicitly.

While there, promote reduce intents from "initial implementation"
and indicate that they are available for coforalls too.