Add four tests of sending initialized fields to functions in Phase 1
The behavior today is inconsistent - if the return value of the function being
called is used as the initial value of a field, we allow the call to contain a
prior field.  However, that same call is not allowed in other circumstances in
Phase 1.  We would like to be able to send fields that have already been
initialized as arguments to functions during Phase 1 (assuming the intent for
the field does not allow it to be modified, as this would be a double
initialization).