Add a test of functions that are declared as throws with chpldoc
This test failed prior to the previous commit, and now passes.

Covers functions that are declared as throws with various combinations of having
a body, an argument, a return type, and a documentation comment.  Also checks
that being inside or outside an explicit module doesn't have an effect.

Does not cover methods on types, but I have no reason to expect this doesn't
work.