No longer use clang to emit the .o file
We were running into some problems with the clang/llvm integration, where
--llvm-wide-opt would need to change the DataLayout for the module but
the call to clang that we were using to produce a .o file,
EmitBackendOutput, would always reset the DataLyout string, and then if
we tried to change it in a pass within we'd get core dumps.

So, before this commit, we were running LLVM optimization passes twice:
once in our invocation of optimizations, and also once in
EmitBackendOutput.

I was curious about how Swift handles this issue, since it also
integrates to some degree with clang. Looking at the Swift source code,
especially IRGen.cpp, demonstrated an alternative way to do the same
thing using LLVM calls rather than the Clang pipeline.

At the same time, I had a hunch that we weren't setting some important
optimization settings for our own optimization pipeline - and indeed
we weren't adding a TargetTransformInfoWrapper pass, which both clang
and Swift do. While there, I looked at other settings that clang and
Swift set that we do not. The result should hopefully be more robust
and give better optimization.