Add --print-commands output for target features, llvm flags, fix
I noticed that --mllvm wasn't working correctly. I also was uncertain
if the new code added in setupModule was getting the correct target
features. To debug these, added printing of them with --print-commands.

I noticed that we were saving Clang->getTargetOpts() before the
Features part was filled in, but we no longer needed to save it,
so replaced the saving with just accessing it from Clang.

Moved parsing llvm options earlier in to setupClang since at that
point in initialization we can access the Clang -mllvm options too,
and we need to run it before we start running LLVM passes.