Add a workaround for errno for ZMQ to work with --llvm
errno is actually normally a macro (and one that is relatively
complex). Our LLVM macro support isn't up to the task of determining
what errno is. So, to enable ZMQ + --llvm, I'm adding a little
inline function to return errno, which should be sufficient for ZMQ.

Notes:
 - ZMQ also provides zmq_errno, but this is probably not inlined.
 - The new macro, chpl_macro_int_errno, follows the same strategy
   as other similar workarounds (such as chpl_macro_float_isnan,
   chpl_macro_int_EEOF and others). We could generate these workaround
   functions automatically if we had a way to ask Clang for the return
   type of an expression.

I think it's an open question whether or not errno should be available
from a standard Chapel module. I've created issue #7499 to discuss.