Explicitly skip some tests with custom sub_tests for perf testing
The real sub_test knows which tests to skip while running performance testing,
but most custom sub_tests don't. This results in tests running when they
shouldn't be, so add explicit SKIPIFs for them. This is similar to #5948

I noticed this while running release-over-release performance testing because
the default-binary-name test fails for older releases.