Fix sign/zero extending for LLVM GEP instructions
The test test/arrays/elliot/nearMaximalUintArray.chpl was failing with
LLVM compilation because of a bug in how Chapel was generating
getelementptr instructions.

The relevant case is something like this:

    %eptr = getelementptr inbounds i64, i64* %ptr, i8 %idx

According to the LLVM IR manual, such a GEP should sign-extend %idx
to the size of the pointer. But in the case of nearMaximalUintArray.chpl,
we need it to zero extend instead.

Since LLVM IR integer types do not differentiate between signed and
unsigned, solving this problem amounts to adding a sign/zero extend in
the Chapel code generator, where we know whether the index is signed or
not. I accomplished that by creating a new function in codegen/expr.cpp,
extendToPointerSize, and calling it in select places.

In order to prevent similar problems in other places that the Chapel
code generator produces getelementptr instructions, I replaced calls to
IRBuilder::CreateInBoundsGEP with calls to a new function that verifies
that the index of a GEP we are creating is at least intptr_t bits wide.