Stop serializing foralls in parallel/forall/vass/memleaks*
This reverts #2537. That PR quieted some memleaks by wrapping a forall with a
serial stmt. The root cause was never understood, but the sporadic leaks seem
to be gone now, so remove the serial stmts and references to the JIRA issue.

Passes 100,000 trials for test/parallel/forall/vass/memleaks*.chpl with the
quickstart configuration for --local and --no-local.