Fix memory leak caused by "".join(("",""))
This currently results in returning a string with len 0 but buf[]
allocated and _size > 0.  The len being 0 results in isEmptyString()
returning true, so deinit() doesn't free the buf[].

Avoid this by handing a resulting joinedSize 0 as a special case, and
return a real empty string.

This fixes the leak in test/types/string/psahabu/join-tuples.chpl