Adjust cpp-lines test to tolerate environment variable being set
In some testing configurations, the perftools-base module is setting
CHPL_CG_CPP_LINES=1. That was causing failures for noCppLines.

This commit makes this test more robust by passing --no-cpp-lines
which has the effect of overriding the environment variable.

While there, made these tests .noexec since the only useful thing being
tested here is whether or not #line numbers were emitted in the generated C,
and that doesn't require that the program be run.