Add test of align operator
This test shows some of the current behaviors of using the 'align'
operator, some of which seem surprising to me, others less so.  I'm
putting this in to capture some of the current behaviors of 'align' as
part of issue #7290, where they may change if we proceeded with that
proposal.