Move global-to-wide later in LLVM pipeline
I noticed that we were running global-to-wide earlier than I intended;
EP_ScalarOptimizerLate seems too early since it happens before
vectorization. This commit changes it to run as the last stage in
optimization with EP_OptimizerLast (and then we'll run inlining etc again
in a 2nd set of opts).

Also switches to using two Module Pass Managers; one for optimization in
address-space-100 that concludes with global-to-wide, and then a second
to do clean-up optimization on that. These are currently two pass
managers so that it's easier to debug their output.

PR #7542 shows the start of another way to enable saving the LLVM module
for debugging purposes at key points.