Remove initialize method from Timer record (which was empty)
We've decided that we would like to remove as many uses of the old
"initialize()" method (which was used to do some operations after the completion
of the default constructor, without having to redefine the default constructor
to get its functionality) as possible.

This initialize method should be trivial to remove - it had no contents and the
type is not one that has special compiler support, so we should not see an
impact after its removal.