Remove the empty initialize method from this test, and a related comment
Based on the context, I believe the comment within the initialize method was
related to the commented out alternate declaration for the array field.  I
could add an alternative commented out initializer to replace the initialize()
method, but I don't think that is necessary.

The test still passes with this change.