Switch two initialize() methods in tests to initializers
These two initialize() methods overwrite the value of an array field that is
dependent on a previous domain field.  Based on the way they were written and
the other variations of the same name, I am guessing that these methods being
initialize() methods instead of constructors wasn't important, so I am changing
them to initializers.

In more detail, test3.chpl sets the array field to the result of a call that
takes the domain as an argument and returns an array over that domain where the
contents are initialized to 5.0.  I decided to make this a Phase 1 initializer,
making the initial contents of the array set to that value and relying on
omitted initialization for the domain field.  If Phase 1 were the default, we
would not need to write a super.init() for this to be a Phase 1 initializer.

test6.chpl sets the array field to the result of a call which doesn't care about
the domain field.  I decided to leave that as a Phase 2 initializer, relying
on omitted initialization for both fields during Phase 1.  Should we make Phase
1 the default, this initializer will only set the contents of the array field
once.

Both of these tests continue to pass after this change.