Replace more initialize() methods with initializers
The initialize methods for the distributions, domains, and arrays defined in
these two files set the value of one field, meaning that it is not very useful
to have it as an argument.  Make their contents the Phase 2 of initializers that
takes in some of the fields.

These initializers would benefit from being able to set generic fields in Phase
2 and in some cases from being able to set const fields in Phase 2