Add chpl_comm_regMemHeapPageSize().
Add chpl_comm_regMemHeapPageSize(), which returns the hugepage size in
the comm layer registered heap, or the system page size if the comm
layer isn't managing a registered heap.  Reduce the return value from
comm=ugni chpl_comm_regMemAllocThreshold() to just the size of a single
hugepage, since after all that's the actual threshold for usefulness.
Move the decision point for a 2-page threshold for array allocations out
of the registered heap from chpl_comm_regMemAllocThreshold() to the
memory layer's array allocation code, since that's the point at which we
actually know we want at least 2 pages for localization purposes.  (This
logic will need to change eventually, because really the test should be
that the size is at least as many heap pages as there are NUMA domains.
But that's work for the future.)  Having done that, adjust the tests in
the chpl_comm_regMem alloc, post-alloc, and free routines so that they
don't apply a threshold test but only check that we're using hugepages,
thus putting the onus on the callers to decide when to use comm layer
allocation.