Stop throwing `--enable-allow-gcc4` for the GASNet build
`--enable-allow-gcc4` allows "the use of a broken gcc/g++ 4.0-4.2 compiler". It
doesn't actually do anything other than quiet a warning/error that gasnet will
throw if it detects you're using a broken compiler.

gcc 4.2 is almost 10 years old at this point, so I think we can stop throwing
it and just prefer that people upgrade gcc versions or quiet the warning
themselves.

This was originally added in e7f7577d40 in 2008, when gcc 4 was new.