Fix sparse valgrind errors; file future against base issue
Last night's valgrind failures on these new tests were not due to the
new behavior the tests were exercising, but rather the use of a
conditional expression to declare a sparse domain's type.  To work
around the issue, I rewrote the tests to avoid the sparse domain (and
reduced the number of compiles from 2 per test down to 1 to save
compile time), and filed a valgrind-only future against the use of the
conditional type expression.  Given that I've recently been looking at
memory leaks related to dmapped expressions, I'm wondering whether
they could be related.