Add two futures for initializers discovered when making our report
The first shows that having an initializer accidentally causes virtualization to
behave incorrectly.  The second demonstrates that a generic record with a fully
specified declaration for an instance will fail if no initializer with an
argument per generic field is defined, even though the instance would be given
the result of a direct call that does work.