Add support for plain on statements around field initialization
In #6774, we discussed whether it was appropriate to allow on statements around
field initialization and super.init()/this.init() calls in initializers, because
they do not insert parallelism in their own right and so would not violate the
Phase 1 initializer rules.  We decided to allow them around field initialization
for this reason, but not around super.init() and this.init() calls because it is
better to have the Phase division as visible as possible in the body of the
function.

This commit implements that decision.  It continues to track the presence of on
blocks in the body, but will only generate an error message for the super.init()
and this.init() calls.  If a field initialization is in an on block, it will
check if the block surrounding the on block is allowed to contain field
initialization (so that on blocks in loops, etc. will not hide that the field
initialization is improper).  This meant adding tracking for outer blocks.