Testing updates for the new on block behavior with initializers
Changes the expected output for hasCobeginOnFieldInit to be consistent in all
configurations, and the expected output for hasOnFieldInit based on the program
now being allowed to compile and run.  Adds a test, deeperOn, to verify that
having a nested block statement between the inner on statement and the cobegin
does not prevent us from getting the cobegin's error message.  Also, removes
hasOnFieldInitLocal.future and replace it with parentCall/inOnLocal (as the
output for field initialization no longer varies depending on whether the
compilation is local or no-local/comm=gasnet, etc., but the output for
super.init() and this.init() still does).