Re-enable hpl.hpcc2012 testing
It seems that the fixes to the align operator fixed this test?  Re-enable
it to regain gasnet testing and performance testing.  If it continues to
generate thrash in the future, this PR could be reversed, or we could
decide to take it off life support.  I'm curious whether it will go back
to being as stable as it was before, though.