Quiet test_record_of_array_type future under valgrind (#7622)
[reviewed by @ronawho]

Similar to #7266, for some reason the first line of output for this test is occasionally missing the first character. This happens to the third line as well, which the prediff already filters out. Instead of investigating this further, just update the prediff to only grab the second line of output.