Fix the post-parse check for yields and returns from module scope
The checkModule function was walking the statements in the module's body
looking for yields and returns to issue errors about.  However, it didn't
dive into any statements, so skipped over whole BlockStmts without looking
inside of them.

Update it to look at all of the CallExprs in the module body that have the
module itself as a parentSymbol.

Remove a .future that yielded from the module level.