Avoid overwriting enum documentation comment with comment for first constant
Alters the processing of enum declarations so that we can grab the enum
documentation comment before it gets overwritten by the documentation comment
for the first enum constant.  This meant moving the creation of the EnumType to
the processing of the TENUM token; changing the return type of enum_ls to be
a vector of DefExprs; and traversing that vector to determine the default value
for the type, and to get the constants to refer to the overall type itself.

While it means building and traversing a vector instead of inserting the
contents into the EnumType as we encounter them, this seems like a better design
for the change than my other attempt which had made the value of the new
enum_header rule unimportant.

Handles the bug portion of #5022.  It should be pretty easy to add the feature
portion, but I decided to make that a separate action.

Allows test/chpldoc/enum/docConstants.doc.chpl to pass.