Fix bug with const-in tuple formals under multilocale configs
A tuple formal with the 'const in' intent was left with an 'in' intent
by resolveIntents. A program can be written such that the actual
corresponding to the formal was a wide reference. For a normal record
this would not be a problem, as codegen would transform this pattern
such that the actual was dereferenced before the function call.

Tuples are different because codegen transforms tuple formals into
references (as decided by the 'argMustUseCPtr' function). This means
that we'll end up passing a wide-reference to a narrow reference, which
is currently unhandled by codegen.

The more principled solution would be to adjust resolveIntents to treat
'in' and 'const in' more similarly, but that seems challenging to me at
this time. The easiest fix, implemented by this commit, is to modify
'insertWideReferences' to insert a PRIM_DEREF when it sees this pattern.