Enable enums to coerce to uints
I needed to adjust some of the ChapelBase modification done in r21284.

In particular, we have the following param pair to support uint/int
arithmetic:

  proc +(a: uint(64), param b: uint(64)) ...
  proc +(param a: uint(64), b: uint(64))

The problem is that resolution rules for "param prefers" indicate that
we should prefer a function with a param argument to one without,
even if the types don't match. That inteferes with the preference
of int(64) over uint(64) when converting from enums.