Fix issues with using enums in generic functions
Resolves #5414.

Previously, we experienced an issue with enums declared within the scope of a
generic function - when the function was instantiated, we would fail to find the
enum constants.

As it turns out, our copy function on EnumType AST nodes was broken and we
hadn't noticed because we never defined EnumTypes within generic functions.  It
was broken in that it never set the enum constants to have the type of their
defining enum, and it never created a version of all the compiler-generated enum
default functions for the new copy of the EnumType.

This change makes the generation of the enum constant copies use the new type.
It also makes the generation of the enum default functions into something that
can be called outside of buildDefaultFunctions, and calls it when a copy of the
type is created.  Finally, it tags the generated functions as generic when it
creates them, since creating them after the start of functionResolution means
that they avoid the tagIfGeneric iteration through all functions - this allows
us to appropriately match against them during resolution.

Allows test/types/enum/scope.chpl and test/types/enum/scope-type.chpl to pass,
as well as test/types/enum/scope-also-generic.chpl (which was written as part of
this work).