Testing updates for the enum in generic functions bug fix
The .good for scope.chpl and scope-type.chpl was expecting the numerical value,
but it should have been expecting the string version of the constant.  Fix that.

Also adds two new variations: scope-also-generic.chpl (for generic functions due
to having no type on an argument), and scope-works.chpl (which is the concrete
function case and worked prior to this change).  All four now pass.