Add two additional tests that work
I decided to verify that the other default functions were also able to work with
this change.  As it turns out, only assignment and iteration are fine, the enum
first function and the enum size function do not work currently.  Commit the two
that do work, while I investigate the other two to see if I can tack them on?